rename(VJNP-57): Rename components directory AnswererSelector -> SubjectSelection
